+++ code/src/cljs/sixsq/nuvla/ui/credentials/spec.cljs
@@ -55,21 +55,22 @@
 ; Keypair
 
 (s/def ::public-key utils-spec/nonblank-string)
-(s/def ::private-key string?)
+(s/def ::public-key-optional (s/nilable string?))
+(s/def ::private-key-optional (s/nilable string?))
 
 ; SSH
 
 (s/def ::ssh-credential (s/keys :req-un [::name
                                          ::description]
-                                :opt-un [::public-key
-                                         ::private-key]))
+                                :opt-un [::public-key-optional
+                                         ::private-key-optional]))
 
 ; GPG
 
 (s/def ::gpg-credential (s/keys :req-un [::name
                                          ::description
                                          ::public-key]
-                                :opt-un [::private-key]))
+                                :opt-un [::private-key-optional]))
 
 ; MinIO