+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Added
 
   - Profile - Group tab
+  - Support for notification of creation of data-record, used for the
+    new Blackbox application
+  - New layout of data-records and data-sets to better understand and view
+    data-records
 
 ### Changed