+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Updated SixSq legal status (Sàrl to SA)
   - Acl widget - use peers instead of search user resource
   - Session - get peers support
+  - Fix fields validation for apps during editing 
 
 ## [2.19.1] - 2021-06-24