+++ CHANGELOG.md
@@ -5,9 +5,13 @@
 ### Added
 
   - Profile - Group tab
-
+  
 ### Changed
 
+  - Edge Detail - Improve warning and error messages on NB update
+  - Edge - Search icon under input fix
+  - Edge - Enhance responsiveness of statistics
+  - Edge - Regression fix, double anchor to show additional filters
   - Updated SixSq legal status (Sàrl to SA)
   - Acl widget - use peers instead of search user resource
   - Session - get peers support