+++ CHANGELOG.md
@@ -4,14 +4,21 @@
 
 ### Added
 
+  - Apps with environment variables named "S3_CRED" or "GPG_CRED" automatically
+    show a dropdown of available credentials of that type in the launch modal.
   - Profile - Group tab
 
 ### Changed
 
+  - Fixed local validation for creating new credentials (GPG and SSH)
+  - Fix fields validation for apps during editing
+  - Edge Detail - Improve warning and error messages on NB update
+  - Edge - Search icon under input fix
+  - Edge - Enhance responsiveness of statistics
+  - Edge - Regression fix, double anchor to show additional filters
   - Updated SixSq legal status (Sàrl to SA)
   - Acl widget - use peers instead of search user resource
-  - Session - get peers support
-  - Fix fields validation for apps during editing 
+  - Session - get peers support 
 
 ## [2.19.1] - 2021-06-24